Add a central Settings page (top nav) with a sub-nav
Settings were scattered across three sections (Niagawan automation, Payroll
items, Attendance shifts). Add a single top-level Settings in the nav bar (admin
only) with a sub nav bar switching between Automation / Payroll items / Attendance.

Each panel is extracted into a shared component (components/settings/*) so there's
no duplicated logic — the existing per-section settings pages now render the same
components, so those routes keep working unchanged.
